Introduction gives overview of chapter
This new introduction flags the transition from "functions that compute" to "data types of evidence", promising to bring the ideas together in the Decidability chapter.

For context, I spent the first half of this chapter puzzled about why we weren't just defining `<=` as we had `+` or `*`.  

The section on "Decidability" is removed as it is now no longer required.